+++ src/com/inventory/Salles/SalleInfo.java
@@ -276,7 +276,7 @@ public class SalleInfo extends javax.swing.JDialog {
             Double sellPrice = Double.valueOf(listSalles.getValueAt(i, 2).toString());
 //            Double totalRevenue = sellPrice;
 
-            Double totalRevenue = sellPrice * Integer.valueOf(listSalles.getValueAt(i, 3).toString());
+            Double totalRevenue = sellPrice * Double.valueOf(listSalles.getValueAt(i, 3).toString());
             totalPaye = totalPaye + totalRevenue;
  txtTotal.setText(totalPaye.toString());
         }
